Crate oak_baremetal_launcher

Modules

instance ğŸ”’

instance_crosvm ğŸ”’

instance_qemu ğŸ”’

lookup ğŸ”’

server ğŸ”’

Structs

Args ğŸ”’

BaremetalCommunicationChannel

Singleton responsible for sending requests, and receiving responses over the underlying communication channel with the baremetal runtime.

Client

Client implementation that communicates with the communication task via a bmrng channel. Used by various parts of the loader to communicate with the trusted runtime.

VmParams

Parameters used for launching VM instances

Enums

Mode ğŸ”’

Functions

main ğŸ”’

path_exists ğŸ”’